Move to common spark file; change format of signal
